Extract flood fill algorithm from FillTilemapTool
