+++ Source/Plugins/Tilemaps/Core/Tilemaps.Core.csproj
@@ -48,6 +48,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Tilemaps\Components\TilemapChangedEventArgs.cs" />
     <Compile Include="Tilemaps\Components\TilemapCollisionSource.cs" />
+    <Compile Include="Tilemaps\TilemapFloodFill.cs" />
     <Compile Include="Tilesets\TileCollisionShapes.cs" />
     <Compile Include="Tilesets\TileCollisionShape.cs" />
     <Compile Include="Tilesets\TilesetAutoTileItem.cs" />
